Move stuff from formula to prop
