+++ engine/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod formula;
+pub mod prop;
 
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
