+++ engine/src/main.rs
@@ -1,6 +1,6 @@
-use deducktion_engine::formula::parser::settings::ParsingSettings;
+use deducktion_engine::prop::parser::settings::ParsingSettings;
 
 fn main() {
     let settings = ParsingSettings::default();
-    deducktion_engine::formula::test_parser("not (not A or not B)", &settings);
+    deducktion_engine::prop::test_parser("not (not A or not B)", &settings);
 }
