+++ engine/src/prop/mod.rs
@@ -1,5 +1,6 @@
-//! Formulas of propositional logic
+pub use formula::Formula;
 
+mod formula;
 pub mod parser;
 
 pub fn test_lexer(s: &str, settings: &parser::settings::ParsingSettings) {
@@ -17,30 +18,4 @@ pub fn test_parser(s: &str, settings: &parser::settings::ParsingSettings) {
         .parse(tokens.as_slice())
         .unwrap();
     dbg!(formula);
-}
-
-#[derive(Debug, Clone)]
-pub enum Formula {
-    Var(String),
-    Unary {
-        connective: UnaryConnective,
-        arg: Box<Formula>,
-    },
-    Binary {
-        connective: BinaryConnective,
-        lhs: Box<Formula>,
-        rhs: Box<Formula>,
-    },
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum UnaryConnective {
-    Not,
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum BinaryConnective {
-    And,
-    Or,
-    Imp,
 }
